+++ docs/CHANGELOG.md
@@ -18,6 +18,7 @@ VoxConvert:
    - Added option to keep the input file palette and don't perform quantization
    - Allow to export the palette to png
    - Allow to generate models from heightmap images
+   - Allow to run lua scripts to modify volumes
 
 Thumbnailer: